# 06_IMPLEMENTATION/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Marpex CRM API
  version: 0.1.0
  description: |
    Základná OpenAPI špecifikácia pre MVP verziu Marpex CRM.
servers:
  - url: http://localhost:3005
    description: Local API
...
paths:
...
  /api/customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: segment
          schema:
            type: string
            enum: [oem, vyroba, integrator, servis, other]
        - in: query
          name: category
          schema:
            type: string
            enum: [A, B, C]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Customer list
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Page:
              schema:
                type: integer
            X-Limit:
              schema:
                type: integer
            X-Total-Pages:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, segment]
              properties:
                name:
                  type: string
                  example: Demo zákazník a.s.
                segment:
                  type: string
                  enum: [oem, vyroba, integrator, servis, other]
                strategicCategory:
                  type: string
                  enum: [A, B, C]
      responses:
        '201':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ApiError'
...
components:
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
      description: Voliteľné číslo stránky.
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Voliteľný page size.
  responses:
    ApiError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validation:
              value:
                error: Validation error
                code: VALIDATION_ERROR
                details:
                  field:
                    - Povinné pole
  schemas:
    ApiError:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        segment:
          type: string
        strategicCategory:
          type: string
          nullable: true
...